(refactor) mvc 패턴 기준으로 로직 분리
