+++ src/main/java/racingcar/Application.java
@@ -1,17 +1,12 @@
 package racingcar;
 
-import java.util.List;
+import racingcar.controller.GameController;
+import racingcar.service.GameService;
 
 public class Application {
     public static void main(String[] args) {
         // TODO: 프로그램 구현
-        GameManager gameManager = new GameManager();
-        gameManager.init();
-
-        while (!gameManager.play()){
-            System.out.println();
-        }
-
-        gameManager.gameEnd();
+        GameController controller = new GameController();
+        controller.playGame();
     }
 }
