+++ src/main/java/racingcar/controller/GameController.java
@@ -1,0 +1,22 @@
+package racingcar.controller;
+
+import racingcar.service.GameService;
+
+public class GameController {
+
+    private final GameService gameService;
+
+    public GameController(){
+        gameService = new GameService();
+    }
+
+    public void playGame(){
+        gameService.init();
+
+        while (!gameService.play()){
+            System.out.println();
+        }
+
+        gameService.gameEnd();
+    }
+}
